in preparation for warnings cleanup to solve Lucene 2.9 deprecations: fixing typos, unused imports, unimportant glitches found by findbugs
git-svn-id: https://svn.jboss.org/repos/hibernate/search/trunk@18710 1b8cb986-b30d-0410-93ca-fae66ebed9b2

+++ src/main/java/org/hibernate/search/impl/SimpleIndexingProgressMonitor.java
@@ -72,8 +72,8 @@ public class SimpleIndexingProgressMonitor implements MassIndexerProgressMonitor
 	protected void printStatusMessage(long starttimems, long totalTodoCount, long doneCount) {
 		long elapsedMs = System.currentTimeMillis() - starttimems;
 		log.info( "{} documents indexed in {} ms", doneCount, elapsedMs );
-		float estimateSpeed = (float) doneCount * 1000f / elapsedMs ;
-		float estimatePercentileComplete = (float) doneCount * 100f / (float) totalTodoCount ;
+		float estimateSpeed = doneCount * 1000f / elapsedMs ;
+		float estimatePercentileComplete = doneCount * 100f / totalTodoCount ;
 		log.info( "Indexing speed: {} documents/second; progress: {}%", estimateSpeed, estimatePercentileComplete );
 	}
